requirements.txt
streamlit
torch
torchvision
transformers
Pillow
opencv-python
numpy
pandas
tqdm
ddgs
icrawler
requests
kaggle
